backend/requirements.txt
Flask>=2.2.2
Flask_Cors==3.0.10
gensim
matplotlib
numpy
pandas
Pyrebase4==4.6.0
scikit_learn
spacy
Werkzeug
Flask-MySQLdb
